Changed tables data

+++ src/main/resources/db/migration/V20250518233745__create_staff_table.sql
@@ -2,7 +2,7 @@ CREATE TABLE staff (
                        id SERIAL PRIMARY KEY,
                        user_id INT UNIQUE NOT NULL REFERENCES users(id) ON DELETE CASCADE,
                        hire_date DATE NOT NULL,
-                       salary BIGINT NOT NULL,
+                       salary INTEGER NOT NULL,
                        working_days VARCHAR(100) NOT NULL,
                        shift_start TIME NOT NULL,
                        shift_end TIME NOT NULL
